% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{getAveragingTransform}
\alias{getAveragingTransform}
\title{get function that calculates new feature: average of <avg_col> by group}
\usage{
getAveragingTransform(group_cols, avg_col, out_col, fit_dt, n_threshold = 500)
}
\arguments{
\item{group_cols:}{vector of column names to group by the average calculation}

\item{avg_col:}{column to average}

\item{output_col:}{column name to output the new feature}

\item{fit_dt:}{data.table with columns in group_cols and avg_col}

\item{n_threshold:}{calculate average only when group size exceed this}
}
\value{
list(col, fun) where fun is the transformation function
}
\description{
get function that calculates new feature: average of <avg_col> by group
}
