% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{getCategTransform}
\alias{getCategTransform}
\title{get function that encodes away less common values in a feature
saves you from those errors on missing factor levels in test set}
\usage{
getCategTransform(col, fit_dt, threshold = 0, default_value = "other")
}
\arguments{
\item{col:}{transformation function will transform this column}

\item{fit_dt:}{use data in fit_dt to calculate less common feature values}

\item{threshold:}{retain only those values that occur more than this many times in fit_dt}
}
\value{
list(col, fun) where fun is the transformation function
}
\description{
get function that encodes away less common values in a feature
saves you from those errors on missing factor levels in test set
}
