% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{getDiscretizeTransform}
\alias{getDiscretizeTransform}
\title{get function that discretizes numerical feature to factor}
\usage{
getDiscretizeTransform(col, fit_dt, n = 10)
}
\arguments{
\item{col:}{transformation function will transform this column}

\item{fit_dt:}{use data in fit_dt to calculate discretization boundaries}

\item{n:}{calculate this many discretization boundary points}
}
\value{
list(col, fun) where fun is the transformation function
}
\description{
get function that discretizes numerical feature to factor
}
