% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{processNAColumn}
\alias{processNAColumn}
\title{set NA values in column col to 0 and add new column that is 0/1 NA indicator}
\usage{
processNAColumn(input_dt, col)
}
\arguments{
\item{input_dt:}{data.table with column <col>}

\item{col:}{column name in input_dt, is of numeric type}
}
\value{
TRUE, modifies input_dt by reference
}
\description{
set NA values in column col to 0 and add new column that is 0/1 NA indicator
}
